rename package to codehost_scenario
