+++ dev/codehost_scenario/config/BUILD.bazel
@@ -3,6 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "config",
     srcs = ["config.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/dev/tst/config",
+    importpath = "github.com/sourcegraph/sourcegraph/dev/codehost_scenario/config",
     visibility = ["//visibility:public"],
 )
