+++ dev/codehost_scenario/github_client.go
@@ -1,4 +1,4 @@
-package tst
+package codehost_scenario
 
 import (
 	"context"
@@ -10,7 +10,7 @@ import (
 	"github.com/google/go-github/v53/github"
 	"golang.org/x/oauth2"
 
-	"github.com/sourcegraph/sourcegraph/dev/tst/config"
+	"github.com/sourcegraph/sourcegraph/dev/codehost_scenario/config"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
